refactor: rename usr_local_bin to src
