+++ src/incus-app-container-files/get-install-script.ts
