+++ src/incus-app-container-files/read-from-url.ts
