+++ src/incus-app-container-files/read-text-files.ts
