+++ src/incus-app-container-files/status-spinner-resource.ts
