+++ src/incus-app-container-files/template/alpine-319-cloud-install
