+++ src/incus-app-container-files/type-guard.ts
