+++ test/multi-argument.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "https://deno.land/std@0.220.1/testing/bdd.ts";
 import { assertEquals } from "https://deno.land/std@0.220.1/assert/assert_equals.ts";
-import { resolveMultiArgument } from "../usr_local_bin/incus-app-container-files/multi-argument.ts";
+import { resolveMultiArgument } from "../src/incus-app-container-files/multi-argument.ts";
 
 describe("multi-argument", () => {
   describe("resolveMultiArgument", () => {
