+++ test/ssh-key.test.ts
@@ -1,7 +1,5 @@
 import { describe, it } from "https://deno.land/std@0.220.1/testing/bdd.ts";
-import {
-  resolveSshKey,
-} from "../usr_local_bin/incus-app-container-files/ssh-key.ts";
+import { resolveSshKey } from "../src/incus-app-container-files/ssh-key.ts";
 
 describe("ssh-key", () => {
   describe("resolveSshKey", () => {
